work in progress
Former-commit-id: cf5277200ed33655c4e95eefad050173f443dcf1
Former-commit-id: de9eece0a858756261a757456026fb3d2b3a979f

+++ src/socket_util.h
@@ -178,9 +178,9 @@ typedef struct {
     uint8_t weapon_id;
 
     /*
-     * THING_STATE_BIT_SHIFT_EXT1_OWNER_ID_PRESENT
+     * THING_STATE_BIT_SHIFT_EXT1_SHIELD_ID_PRESENT
      */
-    uint16_t owner_id;
+    uint8_t owner_id;
 
     /*
      * THING_STATE_BIT_SHIFT_EXT1_EFFECT_PRESENT
